+++ util-init.R
@@ -16,6 +16,7 @@
 ## Copyright 2017 by Raphael Nömmer <noemmer@fim.uni-passau.de>
 ## Copyright 2017 by Sofie Kemper <kemperso@fim.uni-passau.de>
 ## Copyright 2017 by Felix Prasse <prassefe@fim.uni-passau.de>
+## Copyright 2019 by Klara Schlüter <schluete@fim.uni-passau.de>
 ## All Rights Reserved.