+++ NEWS.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Add a parameter `editor.definition` to the function `add.vertex.attribute.artifact.editor.count` which can be used to define, if author or committer or both count as editors when computing the attribute values. (#92, ff1e147ba563b2d71f8228afd49492a315a5ad48)
+- Add a new file util-evaluation-plot.R containing functions to plot commit edit types per author and project.
 
 
 ## 3.5